+++ packages/styleguide/stories/Organisms/examples/index.ts
@@ -1,0 +1,2 @@
+export * from './connectedForm';
+export * from './gridForm';
